Refactor InputParam class to delegate parameter creation to InputParamFactory, enhancing code organization and maintainability
